Type filter specification, initial enabled state
The type filter can now disable certain types by default according to
a specification profile in types.json. These events are only gathered
when the user enables them in the filter. This increases performance and
reduces load times.

Improve the type filter style to make it a real legend and hint the user
what they can do with it.